# docs/swagger.yaml
definitions:
...
  request.AppReq:
    properties:
      describe:
        type: string
      host:
        type: string
      icon:
        type: string
      name:
        type: string
      poolId:
        type: integer
      tag:
        type: string
    required:
    - describe
    - host
    - icon
    - name
    - tag
    type: object
  request.AppSecretReq:
    properties:
      accessExpire:
        type: integer
      describe:
        type: string
      refreshExpire:
        type: integer
      scope:
        items:
          type: string
        type: array
    type: object
  request.ProviderReq:
    properties:
      agentId:
        type: string
      clientId:
        type: string
      clientSecret:
        type: string
      type:
        type: string
    type: object
  request.RedirectUriReq:
    properties:
      uri:
        type: string
    type: object
...
info:
  contact: {}
  license:
    name: MIT
    url: https://github.com/hello-jiangxiaoyu/QuickAuth/blob/main/LICENSE
  title: Quick Auth 接口文档
  version: "1.0"
paths:
...
  /api/quick/apps:
    get:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      responses:
        "200":
          description: OK
      summary: list apps
      tags:
      - app
    post:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: body
        in: body
        name: bd
        required: true
        schema:
          $ref: '#/definitions/request.AppReq'
      responses:
        "200":
          description: OK
      summary: create app
      tags:
      - app
  /api/quick/apps/{appId}:
    delete:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: app id
        in: path
        name: appId
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: delete app
      tags:
      - app
    get:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: app id
        in: path
        name: appId
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: get app
      tags:
      - app
    put:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: app id
        in: path
        name: appId
        required: true
        type: string
      - description: body
        in: body
        name: bd
        required: true
        schema:
          $ref: '#/definitions/request.AppReq'
      responses:
        "200":
          description: OK
      summary: modify app
      tags:
      - app
  /api/quick/apps/{appId}/secrets:
    get:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: app id
        in: path
        name: appId
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: list app secret
      tags:
      - app
    post:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: app id
        in: path
        name: appId
        required: true
        type: string
      - description: body
        in: body
        name: bd
        required: true
        schema:
          $ref: '#/definitions/request.AppSecretReq'
      responses:
        "200":
          description: OK
      summary: create app secret
      tags:
      - app
  /api/quick/apps/{appId}/secrets/{secretId}:
    delete:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: app id
        in: path
        name: appId
        required: true
        type: string
      - description: secret id
        in: path
        name: secretId
        required: true
        type: integer
      responses:
        "200":
          description: OK
      summary: delete app secret
      tags:
      - app
    put:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: app id
        in: path
        name: appId
        required: true
        type: string
      - description: secret id
        in: path
        name: secretId
        required: true
        type: integer
      - description: body
        in: body
        name: bd
        required: true
        schema:
          $ref: '#/definitions/request.AppSecretReq'
      responses:
        "200":
          description: OK
      summary: create app secret
      tags:
      - app
...
  /api/quick/providers:
    get:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: tenant host
        in: header
        name: vhost
        type: string
      responses:
        "200":
          description: OK
      summary: list provider info
      tags:
      - provider
    post:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: tenant host
        in: header
        name: vhost
        type: string
      - description: body
        in: body
        name: bd
        required: true
        schema:
          $ref: '#/definitions/request.ProviderReq'
      responses:
        "200":
          description: OK
      summary: create provider
      tags:
      - provider
  /api/quick/providers/{providerId}:
    delete:
      parameters:
      - description: provider id
        in: path
        name: providerId
        required: true
        type: integer
      - description: tenant host
        in: header
        name: vhost
        type: string
      responses:
        "200":
          description: OK
      summary: delete provider
      tags:
      - provider
    get:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: tenant host
        in: header
        name: vhost
        type: string
      - description: provider id
        in: path
        name: providerId
        required: true
        type: integer
      responses:
        "200":
          description: OK
      summary: get provider details
      tags:
      - provider
    put:
      parameters:
      - description: provider id
        in: path
        name: providerId
        required: true
        type: integer
      - description: tenant host
        in: header
        name: vhost
        type: string
      - description: body
        in: body
        name: bd
        required: true
        schema:
          $ref: '#/definitions/request.ProviderReq'
      responses:
        "200":
          description: OK
      summary: modify provider
      tags:
      - provider
  /api/quick/redirect-uri:
    get:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: tenant host
        in: header
        name: vhost
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      summary: get redirect uri list
      tags:
      - tenant
    post:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: tenant host
        in: header
        name: vhost
        type: string
      - description: body
        in: body
        name: bd
        required: true
        schema:
          $ref: '#/definitions/request.RedirectUriReq'
      responses:
        "200":
          description: OK
      summary: create app redirect uri
      tags:
      - tenant
  /api/quick/redirect-uri/{uri}:
    delete:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: tenant host
        in: header
        name: vhost
        type: string
      - description: uri name
        in: path
        name: uri
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: delete app
      tags:
      - tenant
  /api/quick/redirect-uri/{uriId}:
    put:
      parameters:
      - description: user id
        in: header
        name: X-User-ID
        type: string
      - description: user pool id
        in: header
        name: X-Pool-ID
        type: string
      - description: tenant host
        in: header
        name: vhost
        type: string
      - description: uri id
        in: path
        name: uriId
        required: true
        type: string
      - description: body
        in: body
        name: bd
        required: true
        schema:
          $ref: '#/definitions/request.RedirectUriReq'
      responses:
        "200":
          description: OK
      summary: modify app
      tags:
      - tenant
...
securityDefinitions:
  Login:
    in: header
    name: token
    type: apiKey
swagger: "2.0"